+++ internal/provider/provider.go
@@ -311,8 +311,9 @@ func (p *pingfederateProvider) DataSources(_ context.Context) []func() datasourc
 		oauthauthserversettings.NewOauthAuthServerSettingsDataSource,
 		oauthauthserversettingsscopescommonscope.NewOauthAuthServerSettingsScopesCommonScopeDataSource,
 		oauthauthserversettingsscopesexclusivescope.NewOauthAuthServerSettingsScopesExclusiveScopeDataSource,
-		sessionapplicationsessionpolicy.NewSessionApplicationSessionPolicyDataSource,
 		oauthissuer.NewOauthIssuerDataSource,
+		protocolmetadatalifetimesettings.NewProtocolMetadataLifetimeSettingsDataSource,
+		sessionapplicationsessionpolicy.NewSessionApplicationSessionPolicyDataSource,
 		virtualhostnames.NewVirtualHostNamesDataSource,
 	}
 }