+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### FEATURES
 #### Data Sources
 * **New Data Source:** `session-application-session-policy datasource` ([#94](https://github.com/pingidentity/terraform-provider-pingfederate/pull/94))
+* **New Data Source:** `protocol-metadata-lifetime-settings datasource` ([#100](https://github.com/pingidentity/terraform-provider-pingfederate/pull/100))
 
 # v0.2.1 November 10, 2023
 ### BUG FIXES